move StateEstimationRestService
